backend/requirements.txt
fastapi>=0.115.0
mangum>=0.17.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
boto3>=1.34.0
pydantic>=2.11.0
pydantic-settings>=2.5.0
uvicorn>=0.30.0
